+++ pyronyx/src/extensions/nv/ray_tracing.rs
@@ -170,7 +170,7 @@ impl RayTracingDevice for Device {
             (call)(
                 self.handle,
                 acceleration_structure,
-                data.len() as usize,
+                data.len(),
                 data.as_mut_ptr().cast(),
             )
         }
